requirements.txt
#pip>=9.0.1
#Flask==0.12.2
pip
Flask

requests
requests-cache
plotly
#pyopenssl #https
#flask-cassandra
cassandra-driver

flask-login
